requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
sentence-transformers==3.1.1
faiss-cpu==1.8.0
numpy==1.26.4
scikit-learn==1.5.2
pandas==2.2.2
python-dotenv==1.0.1
mlflow==2.16.0
requests==2.32.3
